Add dto and mapper for rental model

+++ front/src/app/components/user-list/user-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { User } from '../../models/user';
 import { UserService } from '../../services/user.service';
+import { AuthService } from '../../services/auth.service';
 
 @Component({
   selector: 'app-user-list',
@@ -12,7 +13,8 @@ export class UserListComponent implements OnInit {
   users: User[];
 
   constructor(
-    private userService: UserService
+    private userService: UserService,
+    private authService: AuthService
   ) { }
 
   ngOnInit() {
@@ -23,7 +25,11 @@ export class UserListComponent implements OnInit {
     this.userService.findAll().subscribe(res => this.users = res);
   }
 
-  delete(id: number){
+  delete(id: number) {
+    if (this.authService.getUsername() === this.users.find(e => e.id === id).username) {
+      alert('You can\'t remove yourself!');
+      return;
+    }
     this.userService.delete(id).subscribe(res => this.fetchData());
   }
